Extract the list type so we can reference it in the unit tests
