+++ database/CMakeLists.txt
@@ -13,6 +13,7 @@ set(Sources
 	CategoryBitfield.cpp CategoryBitfield.h
 	PatchDatabase.cpp PatchDatabase.h
 	PatchFilter.cpp PatchFilter.h
+    PatchListType.h
 )
 
 set(SQLITE_CPP_INCLUDE "${CMAKE_CURRENT_LIST_DIR}/../../third_party/SQLiteCpp/include")
